Adds terms agreement to signup form
Enhances the signup process by requiring users to agree to the Merchant Agreement.

This change introduces a checkbox to the signup form that users must check to proceed. It also includes a link to the Merchant Agreement page for easy access. The submit button is disabled until the agreement is accepted. Also, adds a "Terms" link to the homepage navigation and a "Merchant Agreement" link to the footer.